Update Magnum to version 1.4.0
Added ability to generate decoy sequences on the fly. Decoys adhere to these rules:
1. Enzyme cut sites are held in place and amino acids in between them are reversed. This creates equal numbers of decoy peptides as target peptides.
2. For palindromic peptides, minor steps are taken to shuffle the amino acids. These are not foolproof.
3. Generated decoy proteins take the name given by decoy_filter parameter, plus a 0 or 1, to allow the decoys to be further split for entrapment purposes.

The decoy_filter parameter now takes two values:
1. the identifier to use for decoy sequences
2. a 0 or 1, where 0=do not generate decoys in Magnum, 1=generate decoys in Magnum. This way Magnum functionality from previous versions is preserved if people want to use their own decoy schemes.

If Magnum generates decoy sequences on the fly, these sequences will be appended to a copy of the FASTA database and exported to the output directory for further inspection or use.